+++ __tests__/data/resource.data.ts
@@ -1,0 +1,35 @@
+import { IResource } from '../../src/interfaces/resource.interface';
+
+export const data = <IResource>{
+  decodedBodySize: 89476,
+  dnsLookupTime: 0,
+  encodedBodySize: 89476,
+  fetchUntilResponseEndTime: 286.09999999403954,
+  id: 'v1-1624213181207-9776614057997',
+  initiatorType: 'script',
+  name: 'http://test/js/jquery.min.js',
+  redirectTime: 0,
+  requestStartUntilResponseEndTime: 286.09999999403954,
+  responseTime: 5.5,
+  secureConnectionTime: 0,
+  startUntilResponseEndTime: 286.09999999403954,
+  tcpHandshakeTime: 0.4000000059604645,
+  transferSize: 89787,
+};
+
+export const data2 = <IResource>{
+  decodedBodySize: 12345,
+  dnsLookupTime: 0,
+  encodedBodySize: 12345,
+  fetchUntilResponseEndTime: 274.10000000032051,
+  id: 'v1-1624213181207-173451425222',
+  initiatorType: 'script',
+  name: 'http://test/js/jquery.min.js',
+  redirectTime: 0,
+  requestStartUntilResponseEndTime: 274.10000000032051,
+  responseTime: 4.2,
+  secureConnectionTime: 0,
+  startUntilResponseEndTime: 274.10000000032051,
+  tcpHandshakeTime: 0.3000023094601235,
+  transferSize: 12386,
+};