+++ __tests__/data/vital.data.ts
@@ -1,0 +1,15 @@
+import { IVital } from '../../src/interfaces/vital.interface';
+
+export const data = <IVital>{
+  id: 'v1-1624213181207-9776614057997',
+  name: 'FCP',
+  delta: 848.9000000059605,
+  value: 848.9000000059605,
+};
+
+export const data2 = <IVital>{
+  id: 'v1-1624213181207-173451425222',
+  name: 'TTFB',
+  delta: 320.1000000043205,
+  value: 320.1000000043205,
+};