+++ __tests__/device.spec.ts
@@ -1,7 +1,6 @@
 import { ObjectId } from 'mongoose';
 import { connect, close, clear } from '../__mocks__/database.mock';
-import { IDevice } from '../src/interfaces/device.interface';
-import { IConnection } from '../src/interfaces/connection.interface';
+import { data, data2 } from './data/device.data';
 import DeviceModel from '../src/models/device.model';
 import { DeviceServices } from '../src/services/device.services';
 
@@ -10,21 +9,6 @@ jest.setTimeout(30000);
 beforeAll(async () => await connect());
 afterEach(async () => await clear());
 afterAll(async () => await close());
-
-const data = <IDevice>{
-  id: 'v1-1624213181207-9776614057997',
-  connection: <IConnection>{
-    effectiveType: '4g',
-    rtt: 50,
-    downlink: 8.35,
-  },
-  memory: 8,
-  cpus: 8,
-  url: 'http://test',
-  referrer: '',
-  userAgent:
-    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36',
-};
 
 describe('Database method tests for Device', () => {
   it('should save devices correctly to database', async () => {
@@ -49,6 +33,31 @@ describe('Database method tests for Device', () => {
     data.url = '';
     try {
       await DeviceServices.save(data);
+    } catch (ex) {
+      expect(true).toBe(true);
+      return;
+    } finally {
+      data.url = urlBackup;
+    }
+    expect(false).toBe(true);
+  });
+
+  it('should save multiple devices correctly to database', async () => {
+    const deviceIds: Array<ObjectId> = await DeviceServices.saveMany([
+      data,
+      data2,
+    ]);
+    expect(deviceIds.length).toBe(2);
+
+    const devices = await DeviceModel.find({ _id: { $in: deviceIds } });
+    expect(devices.length).toBe(2);
+  });
+
+  it('should throw an error when wrong data is sent to register multiple devices', async () => {
+    const urlBackup = data.url;
+    data.url = '';
+    try {
+      await DeviceServices.saveMany([data, data2]);
     } catch (ex) {
       expect(true).toBe(true);
       return;