+++ __tests__/vital.spec.ts
@@ -1,6 +1,6 @@
 import { ObjectId } from 'mongoose';
 import { connect, close, clear } from '../__mocks__/database.mock';
-import { IVital } from '../src/interfaces/vital.interface';
+import { data, data2 } from './data/vital.data';
 import VitalModel from '../src/models/vital.model';
 import { VitalServices } from '../src/services/vital.services';
 
@@ -9,13 +9,6 @@ jest.setTimeout(30000);
 beforeAll(async () => await connect());
 afterEach(async () => await clear());
 afterAll(async () => await close());
-
-const data = <IVital>{
-  id: 'v1-1624213181207-9776614057997',
-  name: 'FCP',
-  delta: 848.9000000059605,
-  value: 848.9000000059605,
-};
 
 describe('Database method tests for Vital', () => {
   it('should save web vitals correctly to database', async () => {
@@ -38,6 +31,31 @@ describe('Database method tests for Vital', () => {
     data.name = '';
     try {
       await VitalServices.save(data);
+    } catch (ex) {
+      expect(true).toBe(true);
+      return;
+    } finally {
+      data.name = nameBackup;
+    }
+    expect(false).toBe(true);
+  });
+
+  it('should save multiple web-vitals correctly to database', async () => {
+    const vitalIds: Array<ObjectId> = await VitalServices.saveMany([
+      data,
+      data2,
+    ]);
+    expect(vitalIds.length).toBe(2);
+
+    const vitals = await VitalModel.find({ _id: { $in: vitalIds } });
+    expect(vitals.length).toBe(2);
+  });
+
+  it('should throw an error when wrong data is sent to register multiple vitals', async () => {
+    const nameBackup = data.name;
+    data.name = '';
+    try {
+      await VitalServices.saveMany([data, data2]);
     } catch (ex) {
       expect(true).toBe(true);
       return;