+++ src/services/device.services.ts
@@ -1,4 +1,5 @@
 import { ObjectId } from 'mongoose';
+import { IDevice } from '../interfaces/device.interface';
 import Device from '../models/device.model';
 
 /**
@@ -19,6 +20,23 @@ export class DeviceServices {
     try {
       await newDevice.save();
       return newDevice._id;
+    } catch (ex) {
+      throw new Error(ex);
+    }
+  }
+
+  /**
+   * save a list of devices into database
+   * @param {Array<IDevice>} data - array of IDevice objects
+   * @returns {Promise<Array<ObjectId>>}
+   * @throws {Error}
+   * @async
+   * @static
+   */
+  public static async saveMany(data: Array<IDevice>): Promise<Array<ObjectId>> {
+    try {
+      const result = await Device.insertMany(data);
+      return result.map((item) => item._id);
     } catch (ex) {
       throw new Error(ex);
     }