+++ src/services/resource.services.ts
@@ -1,4 +1,5 @@
 import { ObjectId } from 'mongoose';
+import { IResource } from '../interfaces/resource.interface';
 import Resource from '../models/resource.model';
 
 /**
@@ -19,6 +20,25 @@ export class ResourceServices {
     try {
       await newResource.save();
       return newResource._id;
+    } catch (ex) {
+      throw new Error(ex);
+    }
+  }
+
+  /**
+   * save a list of resources into database
+   * @param {Array<IResource>} data - array of IDevice objects
+   * @returns {Promise<Array<ObjectId>>}
+   * @throws {Error}
+   * @async
+   * @static
+   */
+  public static async saveMany(
+    data: Array<IResource>,
+  ): Promise<Array<ObjectId>> {
+    try {
+      const result = await Resource.insertMany(data);
+      return result.map((item) => item._id);
     } catch (ex) {
       throw new Error(ex);
     }