+++ src/services/vital.services.ts
@@ -1,4 +1,5 @@
 import { ObjectId } from 'mongoose';
+import { IVital } from '../interfaces/vital.interface';
 import Vital from '../models/vital.model';
 
 /**
@@ -19,6 +20,23 @@ export class VitalServices {
     try {
       await newVitalData.save();
       return newVitalData._id;
+    } catch (ex) {
+      throw new Error(ex);
+    }
+  }
+
+  /**
+   * save a list of web-vitals into database
+   * @param {Array<IVital>} data - array of IVital objects
+   * @returns {Promise<Array<ObjectId>>}
+   * @throws {Error}
+   * @async
+   * @static
+   */
+  public static async saveMany(data: Array<IVital>): Promise<Array<ObjectId>> {
+    try {
+      const result = await Vital.insertMany(data);
+      return result.map((item) => item._id);
     } catch (ex) {
       throw new Error(ex);
     }